Refactor Fame to not use anonymous classes.
In Fame we have a concept of classes. The classes represent types in Fame. We can instantiate them but we have 4 constantes:
- Object (root class)
- Number (primitive)
- String (primitive)
- Boolean (primitive)

Those 4 classes were managed via anonymous classes but the code was hard to understand. Here I did a refactoring to replace this system of anonymous classes by real classes. 

We now have a concept of FM3Constant and a concept of FM3Primitive and their subclasses are singletons holding those constants.

Maybe we can do better than singleton later but this is already better than the previous implementation.